feat: 여러 자동차를 포함하는 CarGroup 일급 컬렉션 추가
